requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
aiosqlite==0.20.0
pydantic==2.9.2
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
redis==5.0.8
cryptography==43.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
joblib==1.4.2
python-dotenv==1.0.1